+++ .azure/hpu-tests-fabric.yml
@@ -70,17 +70,17 @@ jobs:
       displayName: 'Check the driver status'
 
     - bash: |
-        python -m pytest -sv tests_fabric/test_hpu_accelerator.py --forked --junitxml=hpu_fabric_accelerator_test-results.xml
+        python -m pytest -sv test_fabric/test_hpu_accelerator.py --forked --junitxml=hpu_fabric_accelerator_test-results.xml
       workingDirectory: tests/
       displayName: 'HPU Fabric Accelerator test'
 
     - bash: |
-        python -m pytest -sv tests_fabric/test_hpu_strategy.py --forked --junitxml=hpu_fabric_strategy_test-results.xml
+        python -m pytest -sv test_fabric/test_hpu_strategy.py --forked --junitxml=hpu_fabric_strategy_test-results.xml
       workingDirectory: tests/
       displayName: 'HPU Fabric Strategy test'
 
     # - bash: |
-    #    python -m pytest -sv tests_fabric/test_hpu_precision.py --forked --junitxml=hpu_fabric_precision_test-results.xml
+    #    python -m pytest -sv test_fabric/test_hpu_precision.py --forked --junitxml=hpu_fabric_precision_test-results.xml
     #   workingDirectory: tests/
     #   displayName: 'HPU Fabric Precision test'
