+++ .azure/hpu-tests-pl.yml
@@ -73,19 +73,19 @@ jobs:
       displayName: 'Check the driver status'
 
     - bash: |
-        python -m pytest -sv tests_pytorch/test_accelerator.py --forked --junitxml=hpu1_test-results.xml
+        python -m pytest -sv test_pytorch/test_accelerator.py --forked --junitxml=hpu1_test-results.xml
       workingDirectory: tests/
       displayName: 'Single card HPU test'
 
     - bash: |
-        python -m pytest -sv tests_pytorch/test_accelerator.py --forked --hpus 8 --junitxml=hpu8_test-results.xml
+        python -m pytest -sv test_pytorch/test_accelerator.py --forked --hpus 8 --junitxml=hpu8_test-results.xml
       workingDirectory: tests/
       displayName: 'Multi card(8) HPU test'
 
     - bash: |
-        python -m pytest -sv tests_pytorch/test_precision.py --hmp-bf16 \
-          'tests_pytorch/ops_bf16.txt' --hmp-fp32 \
-          'tests_pytorch/ops_fp32.txt' --forked \
+        python -m pytest -sv test_pytorch/test_precision.py --hmp-bf16 \
+          'test_pytorch/ops_bf16.txt' --hmp-fp32 \
+          'test_pytorch/ops_fp32.txt' --forked \
           --junitxml=hpu1_precision_test-results.xml
       workingDirectory: tests/
       displayName: 'HPU precision test'
