+++ src/lightning_habana/fabric/plugins/__init__.py
@@ -13,6 +13,6 @@
 # limitations under the License.
 
 from lightning_habana.fabric.plugins.io.hpu_io import HPUCheckpointIO
-from lightning_habana.fabric.plugins.precision.hpu_precision import HPUPrecision
+from lightning_habana.fabric.plugins.precision.precision import HPUPrecision
 
 __all__ = ["HPUPrecision", "HPUCheckpointIO"]
