+++ src/lightning_habana/fabric/strategies/single.py
@@ -22,7 +22,7 @@ from lightning.fabric.utilities.types import _DEVICE, Optimizable
 from torch import Tensor
 from torch.nn import Module
 
-from lightning_habana.fabric.plugins.precision.hpu_precision import HPUPrecision
+from lightning_habana.fabric.plugins.precision.precision import HPUPrecision
 from lightning_habana.fabric.utils.imports import _HPU_AVAILABLE, _TORCH_LESSER_EQUAL_1_13_1
 
 if _HPU_AVAILABLE:
