+++ tests/test_fabric/test_hpu_precision.py
@@ -18,10 +18,10 @@ import torch
 import torch.nn as nn
 from lightning.fabric import Fabric, seed_everything
 
-from lightning_habana.fabric.accelerator.hpu import HPUAccelerator
-from lightning_habana.fabric.plugins.precision.hpu_precision import HPUPrecision
-from lightning_habana.fabric.strategies.hpu_single import SingleHPUStrategy
-from tests.tests_fabric.fabric_helpers import BoringFabric
+from lightning_habana.fabric.accelerator.accelerator import HPUAccelerator
+from lightning_habana.fabric.plugins.precision.precision import HPUPrecision
+from lightning_habana.fabric.strategies.single import SingleHPUStrategy
+from tests.test_fabric.fabric_helpers import BoringFabric
 
 
 class MixedPrecisionModule(nn.Module):
