+++ tests/test_fabric/test_hpu_strategy.py
@@ -11,3 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+import torch
+
+from lightning_habana.fabric.accelerator import HPUAccelerator
+from lightning_habana.fabric.strategies.parallel import HPUParallelStrategy
+from lightning_habana.fabric.strategies.single import SingleHPUStrategy
+
+
+def test_single_device_default_device():
+    assert SingleHPUStrategy().root_device == torch.device("hpu")
+
+
+def test_hpu_parallel_strategy_defaults():
+    strategy = HPUParallelStrategy()
+    assert strategy.process_group_backend == "hccl"
+    assert len(strategy.parallel_devices) == HPUAccelerator.auto_device_count()
