+++ tests/test_pytorch/test_accelerator.py
@@ -27,7 +27,7 @@ elif module_available("pytorch_lightning"):
     from pytorch_lightning.demos.boring_classes import BoringModel
     from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
-from lightning_habana.pytorch.accelerator.hpu import HPUAccelerator
+from lightning_habana.pytorch.accelerator.accelerator import HPUAccelerator
 from lightning_habana.pytorch.plugins.precision import HPUPrecisionPlugin
 from lightning_habana.pytorch.strategies import HPUParallelStrategy, SingleHPUStrategy
 from tests.helpers import ClassifDataModule, ClassificationModel
